Clean up the models
