+++ Models/Match/Blackout.cs
@@ -1,23 +1,12 @@
 namespace CODBO4.Models.Match
 {
-    public class Blackout
+    public class Blackout : Profile.User
     {
-        public class Stats
-        {
-            public int kills { get; set; }
-            public int assists { get; set; }
-            public int deaths { get; set; }
-            public int headshots { get; set; }
-            public int shotsFired { get; set; }
-            public int shotsLanded { get; set; }
-            public int shotsMissed { get; set; }
-        }
-
         public class Entry
         {
             public string identifier { get; set; }
             public string gameMode { get; set; }
-            public object gameModeName { get; set; }
+            public string gameModeName { get; set; }
             public string map { get; set; }
             public string mapName { get; set; }
             public string mapImage { get; set; }
@@ -25,16 +14,12 @@ namespace CODBO4.Models.Match
             public int matchEnd { get; set; }
             public int matchWon { get; set; }
             public int CTS { get; set; }
-            public Stats stats { get; set; }
+            public Common.Stats stats { get; set; }
             public string formatForSite { get; set; }
             public bool privateMatch { get; set; }
         }
 
         public bool success { get; set; }
-        public int uid { get; set; }
-        public string username { get; set; }
-        public string platform { get; set; }
-        public string game { get; set; }
         public string type { get; set; }
         public int matchesCount { get; set; }
         public Entry[] entries { get; set; }
